Add top level playbook and helper script
Ansible has a lot of command line options even if you just want to
do a thing, so I wrote a thing to do the thing. It could be extended
to support more options, but if you support too many options, we just
re-created the underlying interface we're trying to simplify with
a wrapper in the first place.